requirements.txt
numpy
tensorflow
pillow
imageio
pandas
moviepy
matplotlib
requests
